lint

+++ contracts/vm/runtime/LibVMState.sol
@@ -57,7 +57,7 @@ library LibVMState {
     using LibMemorySize for uint256[];
     using LibMemorySize for bytes;
     using LibUint256Array for uint256[];
-    using LibUint256Array for uint;
+    using LibUint256Array for uint256;
     using LibVMState for VMState;
     using LibStackTop for uint256[];
     using LibStackTop for StackTop;
@@ -66,7 +66,9 @@ library LibVMState {
     using LibCast for function(VMState memory, SourceIndex, StackTop)
         view
         returns (StackTop);
-    using LibCast for function(VMState memory, Operand, StackTop) view returns (StackTop)[];
+    using LibCast for function(VMState memory, Operand, StackTop)
+        view
+        returns (StackTop)[];
 
     /// Console log various aspects of the VM state.
     /// Gas intensive and relies on hardhat console so not intended for
@@ -77,15 +79,17 @@ library LibVMState {
         DebugStyle debugStyle_
     ) internal view returns (StackTop) {
         if (debugStyle_ == DebugStyle.Source) {
-            for (uint i_ = 0; i_ < state_.compiledSources.length; i_++) {
+            for (uint256 i_ = 0; i_ < state_.compiledSources.length; i_++) {
                 console.logBytes(state_.compiledSources[i_]);
             }
         } else {
-            uint[] memory array_;
-            uint length_;
+            uint256[] memory array_;
+            uint256 length_;
             if (debugStyle_ == DebugStyle.Stack) {
                 length_ = state_.stackBottom.toIndex(stackTop_);
-                array_ = StackTop.unwrap(stackTop_.down(length_)).copyToNewUint256Array(length_);
+                array_ = StackTop
+                    .unwrap(stackTop_.down(length_))
+                    .copyToNewUint256Array(length_);
             } else if (debugStyle_ == DebugStyle.Constant) {
                 array_ = state_.constantsBottom.down().asUint256Array();
             } else {
@@ -100,10 +104,11 @@ library LibVMState {
         return stackTop_;
     }
 
-    function deserialize(
-        bytes memory serialized_,
-        uint256[] memory context_
-    ) internal pure returns (VMState memory) {
+    function deserialize(bytes memory serialized_, uint256[] memory context_)
+        internal
+        pure
+        returns (VMState memory)
+    {
         unchecked {
             VMState memory state_;
 
@@ -157,7 +162,10 @@ library LibVMState {
     /// Hopefully it goes without saying that the list of pointers MUST NOT be
     /// user defined, otherwise any source can be compiled with a completely
     /// different mapping between opcodes and dispatched functions.
-    function compile(bytes memory source_, uint[] memory pointers_) internal pure {
+    function compile(bytes memory source_, uint256[] memory pointers_)
+        internal
+        pure
+    {
         assembly ("memory-safe") {
             for {
                 let replaceMask_ := 0xFFFF
@@ -166,17 +174,22 @@ library LibVMState {
                 let pointersBottom_ := add(pointers_, 0x20)
                 let cursor_ := add(source_, 2)
                 let end_ := add(source_, sourceLength_)
-            }
-            lt(cursor_, end_)
-            {
+            } lt(cursor_, end_) {
                 cursor_ := add(cursor_, 4)
-            }
-            {
+            } {
                 let data_ := mload(cursor_)
-                mstore(cursor_, or(and(data_, preserveMask_), mload(
-                    add(
-                        pointersBottom_, 
-                        mul(and(data_, replaceMask_), 0x20)))))
+                mstore(
+                    cursor_,
+                    or(
+                        and(data_, preserveMask_),
+                        mload(
+                            add(
+                                pointersBottom_,
+                                mul(and(data_, replaceMask_), 0x20)
+                            )
+                        )
+                    )
+                )
             }
         }
     }
@@ -184,17 +197,17 @@ library LibVMState {
     function serialize(
         StateConfig memory config_,
         uint256 stackLength_,
-                    function(VMState memory, Operand, StackTop)
-                internal
-                view
-                returns (StackTop)[]
-                memory opcodeFunctionPointers_
+        function(VMState memory, Operand, StackTop)
+            internal
+            view
+            returns (StackTop)[]
+            memory opcodeFunctionPointers_
     ) internal pure returns (bytes memory) {
         unchecked {
-            uint size_ = 0;
+            uint256 size_ = 0;
             size_ += stackLength_.size();
             size_ += config_.constants.size();
-            for (uint i_ = 0; i_ < config_.sources.length; i_++) {
+            for (uint256 i_ = 0; i_ < config_.sources.length; i_++) {
                 size_ += config_.sources[i_].size();
             }
             bytes memory serialized_ = new bytes(size_);
